BUG FIX (An assertion is currently unnecessary)
due to the `ne 0,` operation when dealing with short-circuit logic
(land) in `ast.h`, when visiting binary values(instructions) in
`koopair.cpp`, we cannot assert that there is at most one integer
between the two operands.

Again, this is not the perfect way. We can pass some information
to the expression in land, and tells it to return the `ne 0,`
value directly to land. Leave for future work.